- 从数据库中移除设备的在线状态字段。
