+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - 修改现有表结构，添加一些约束条件。
 - 重构用户、角色部分接口的实现，分离角色操作。
 - 更换Apache的MD5计算方法。
+- 从数据库中移除设备的在线状态字段。
 ### Fixed
 - 修复返回结果对象中，loginUser可能为null引起的空指针的问题。
 - 修复请求出现错误时也能不能正确获取用户名称的问题。
